move eslint to different file
